// backend/src/shared/enums.ts
export enum bucketFolders {
	profilePic = 'profilePics',
	marketplace = 'marketplace',
	profileHeader = 'profileHeaders',
	covers = 'covers',
	songs = 'songs',
	invoices = 'invoices',
}

export enum fileTypes {
	jpg = 'jpg',
	jpeg = 'jpeg',
	png = 'png',
	mp3 = 'mp3',
	pdf = 'pdf',
	mp4 = 'mp4'
}

export enum platforms {
	facebook = 'facebook',
	instagram = 'instagram',
	soundcloud = 'soundcloud',
}

export enum profileListItemsTitles {
	Skills = 'Skills',
	ArtistCollaberations = 'Artist Collaberations',
	PreviousReleases = 'Previous Releases',
}